Move JVM and thread functions to a separate module
