+++ jni/jni.cabal
@@ -41,6 +41,7 @@ library
     Foreign.JNI.Unsafe
     Foreign.JNI.Unsafe.Internal
     Foreign.JNI.Unsafe.Internal.Introspection
+    Foreign.JNI.Unsafe.Internal.JVM
   build-depends:
     async >=2.2.2,
     base >=4.14 && <5,
